requirements.txt
appdirs==1.4.4
asgiref==3.7.2
attrs==23.1.0
bcrypt==3.2.0
black==21.5b1
blinker==1.4
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
cryptography==3.4.8
dbus-next==0.2.3
# dbus-python==1.2.18
distro==1.7.0
dj-database-url==2.0.0
Django==3.2.6
django-heroku==0.3.1
django-phonenumber-field==7.1.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==19.6.0
fasteners==0.14.1
flake8==3.9.2
future==0.18.2
gunicorn==20.1.0
httplib2==0.20.2
idna==2.10
importlib-metadata==4.6.4
inflection==0.5.1
iniconfig==2.0.0
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
mccabe==0.6.1
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.1
paramiko==2.9.3
pathspec==0.11.2
pexpect==4.8.0
phonenumbers==8.13.20
Pillow==9.0.1
pluggy==0.13.1
protobuf==3.12.4
# psycopg2==2.9.7
psycopg2-binary==2.9.7
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.2.4
python-dateutil==2.8.2
pytz==2023.3
pyxdg==0.27
PyYAML==6.0.1
regex==2023.8.8
# reportlab==3.6.8
requests==2.25.1
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.4.4
toml==0.10.2
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
whitenoise==6.5.0
zipp==3.16.2